server/requirements.txt
Django==2.0.5
requests
pillow